Move FakeWorkflowsRepo to Spec::Support
