+++ spec/models/manageiq/providers/workflows/automation_manager/configuration_script_source_spec.rb
@@ -1,5 +1,3 @@
-require_relative "fake_workflows_repo"
-
 RSpec.describe ManageIQ::Providers::Workflows::AutomationManager::ConfigurationScriptSource do
   context "with a local repo" do
     let(:ems) { FactoryBot.create(:ems_workflows_automation) }
@@ -19,7 +17,7 @@ RSpec.describe ManageIQ::Providers::Workflows::AutomationManager::ConfigurationS
     before do
       FileUtils.mkdir_p(local_repo)
 
-      repo = FakeWorkflowsRepo.new(local_repo, repo_dir_structure)
+      repo = Spec::Support::FakeWorkflowsRepo.new(local_repo, repo_dir_structure)
       repo.generate
       repo.git_branch_create("other_branch")
 
@@ -212,7 +210,7 @@ RSpec.describe ManageIQ::Providers::Workflows::AutomationManager::ConfigurationS
         end
 
         it "finds all ASL workflows" do
-          FakeWorkflowsRepo.generate(nested_repo, nested_repo_structure)
+          Spec::Support::FakeWorkflowsRepo.generate(nested_repo, nested_repo_structure)
 
           params[:scm_url] = "file://#{nested_repo}"
           record           = build_record
@@ -231,7 +229,7 @@ RSpec.describe ManageIQ::Providers::Workflows::AutomationManager::ConfigurationS
         end
 
         it "finds only ASL workflows" do
-          FakeWorkflowsRepo.generate(well_documented_repo, well_documented_repo_structure)
+          Spec::Support::FakeWorkflowsRepo.generate(well_documented_repo, well_documented_repo_structure)
 
           params[:scm_url] = "file://#{well_documented_repo}"
           record           = build_record
@@ -252,7 +250,7 @@ RSpec.describe ManageIQ::Providers::Workflows::AutomationManager::ConfigurationS
         end
 
         it "finds only ASL workflows" do
-          FakeWorkflowsRepo.generate(hide_and_seek_repo, hide_and_seek_repo_structure)
+          Spec::Support::FakeWorkflowsRepo.generate(hide_and_seek_repo, hide_and_seek_repo_structure)
 
           params[:scm_url] = "file://#{hide_and_seek_repo}"
           record           = build_record
